fixes #20: let ronn build manpages from tarballs
Split the manpage pattern rule into two rules. The first checks for a .git directory to see if it's a git repo, if that rule matches then it uses git to find the last commit date for the ronn source and uses that. Otherwise it falls through to the second rule which just lets ronn default to using the mtime of the source file. This should let ronn build manpages from a tarball.